+++ src/gamspy/formulations/__init__.py
@@ -1,4 +1,12 @@
 import gamspy.formulations.nn as nn
-from gamspy.formulations.nn import Conv2d
+from gamspy.formulations.nn import AvgPool2d, Conv2d, MaxPool2d, MinPool2d
+from gamspy.formulations.shape import flatten_dims
 
-__all__ = ["nn", "Conv2d"]
+__all__ = [
+    "nn",
+    "Conv2d",
+    "MaxPool2d",
+    "MinPool2d",
+    "AvgPool2d",
+    "flatten_dims",
+]